+++ src/Models/Embeddings/CreateRequest.php
@@ -13,7 +13,6 @@ declare(strict_types=1);
 namespace Tectalic\OpenAi\Models\Embeddings;
 
 use Tectalic\OpenAi\Models\AbstractModel;
-use n;
 
 final class CreateRequest extends AbstractModel
 {
@@ -38,10 +37,7 @@ final class CreateRequest extends AbstractModel
     /**
      * Input text to get embeddings for, encoded as a string or array of tokens. To get
      * embeddings for multiple inputs in a single request, pass an array of strings or
-     * array of token arrays. Each input must not exceed 2048 tokens in length.
-     * Unless you are embedding code, we suggest replacing newlines (n) in your input
-     * with a single space, as we have observed inferior results when newlines are
-     * present.
+     * array of token arrays. Each input must not exceed 8192 tokens in length.
      *
      * Example: 'The quick brown fox jumped over the lazy dog'
      *
@@ -50,8 +46,8 @@ final class CreateRequest extends AbstractModel
     public $input;
 
     /**
-     * A unique identifier representing your end-user, which will help OpenAI to
-     * monitor and detect abuse. Learn more.
+     * A unique identifier representing your end-user, which can help OpenAI to monitor
+     * and detect abuse. Learn more.
      *
      * Example: 'user-1234'
      *