+++ src/Models/ImagesGenerations/CreateRequest.php
@@ -72,8 +72,8 @@ final class CreateRequest extends AbstractModel
     public $response_format;
 
     /**
-     * A unique identifier representing your end-user, which will help OpenAI to
-     * monitor and detect abuse. Learn more.
+     * A unique identifier representing your end-user, which can help OpenAI to monitor
+     * and detect abuse. Learn more.
      *
      * Example: 'user-1234'
      *